Clear Kotlin compile warnings surfaced in CI.
Use edge-to-edge system bars, fix annotation targets, coroutines opt-in, and minor test/native cleanups.

+++ app/src/main/java/com/georgv/audioworkstation/engine/NativeEngine.kt
@@ -82,7 +82,7 @@ class NativeEngine @Inject constructor() {
     }
 
     fun playbackLaneLifecycle(laneIndex: Int): PlaybackLaneLifecycle =
-        PlaybackLaneLifecycle.entries[nativeGetPlaybackLaneLifecycle(laneIndex).toInt()]
+        PlaybackLaneLifecycle.entries[nativeGetPlaybackLaneLifecycle(laneIndex)]
 
     fun isPlaybackActive(): Boolean = nativeIsPlaybackActive()
